Rename the project to Softener
I have renamed the project because I want the name to be
GitHub-agnostic. When choosing a name, I was reading about lint (in the
literal sense) and found the following sentence on Wikipedia:

> The accumulation of lint during clothes cleaning can be reduced with
> the use of a fabric softener, which reduces the amount of static
> electricity on clothing surfaces and therefore prevents the lint from
> sticking to the clothes.

I wasn't aware that this is how fabric softener works, but it serves as
a good metaphor.

Unfortunately, the words "soft" and "soften" are already taken as names
on PyPI. However, the "-er" suffix does make more sense as a noun
inflection (i.e., a tool that does something).

Also, the word is cute.

This commit changes the project name and makes a few other minor changes
to prepare it for rerelease as an alpha.